formatted with master

+++ repos/scala/test/disabled/presentation/akka/src/akka/dispatch/MessageHandling.scala
@@ -9,7 +9,12 @@ import akka.event.EventHandler
 import akka.config.Configuration
 import akka.config.Config.TIME_UNIT
 import akka.util.{Duration, Switch, ReentrantGuard}
-import java.util.concurrent.ThreadPoolExecutor.{AbortPolicy, CallerRunsPolicy, DiscardOldestPolicy, DiscardPolicy}
+import java.util.concurrent.ThreadPoolExecutor.{
+  AbortPolicy,
+  CallerRunsPolicy,
+  DiscardOldestPolicy,
+  DiscardPolicy
+}
 import akka.actor._
 
 /**
@@ -19,7 +24,8 @@ final case class MessageInvocation(
     val receiver: ActorRef,
     val message: Any,
     val sender: Option[ActorRef],
-    val senderFuture: Option[CompletableFuture[Any]]) {
+    val senderFuture: Option[CompletableFuture[Any]]
+) {
   if (receiver eq null)
     throw new IllegalArgumentException("Receiver can't be null")
 
@@ -29,24 +35,27 @@ final case class MessageInvocation(
     } catch {
       case e: NullPointerException =>
         throw new ActorInitializationException(
-            "Don't call 'self ! message' in the Actor's constructor (in Scala this means in the body of the class).")
+          "Don't call 'self ! message' in the Actor's constructor (in Scala this means in the body of the class)."
+        )
     }
 }
 
 final case class FutureInvocation[T](
-    future: CompletableFuture[T], function: () => T, cleanup: () => Unit)
-    extends Runnable {
+    future: CompletableFuture[T],
+    function: () => T,
+    cleanup: () => Unit
+) extends Runnable {
   def run = {
     future complete
-    (try {
-          Right(function())
-        } catch {
-          case e =>
-            EventHandler.error(e, this, e.getMessage)
-            Left(e)
-        } finally {
-          cleanup()
-        })
+      (try {
+        Right(function())
+      } catch {
+        case e =>
+          EventHandler.error(e, this, e.getMessage)
+          Left(e)
+      } finally {
+        cleanup()
+      })
   }
 }
 
@@ -80,22 +89,26 @@ trait MessageDispatcher {
   /**
     * Attaches the specified actorRef to this dispatcher
     */
-  final def attach(actorRef: ActorRef): Unit = guard withGuard {
-    register(actorRef)
-  }
+  final def attach(actorRef: ActorRef): Unit =
+    guard withGuard {
+      register(actorRef)
+    }
 
   /**
     * Detaches the specified actorRef from this dispatcher
     */
-  final def detach(actorRef: ActorRef): Unit = guard withGuard {
-    unregister(actorRef)
-  }
+  final def detach(actorRef: ActorRef): Unit =
+    guard withGuard {
+      unregister(actorRef)
+    }
 
-  private[akka] final def dispatchMessage(
-      invocation: MessageInvocation): Unit = dispatch(invocation)
+  private[akka] final def dispatchMessage(invocation: MessageInvocation): Unit =
+    dispatch(invocation)
 
   private[akka] final def dispatchFuture[T](
-      block: () => T, timeout: Long): Future[T] = {
+      block: () => T,
+      timeout: Long
+  ): Future[T] = {
     futures.getAndIncrement()
     try {
       val future = new DefaultCompletableFuture[T](timeout)
@@ -119,14 +132,17 @@ trait MessageDispatcher {
             case UNSCHEDULED =>
               shutdownSchedule = SCHEDULED
               Scheduler.scheduleOnce(
-                  shutdownAction, timeoutMs, TimeUnit.MILLISECONDS)
+                shutdownAction,
+                timeoutMs,
+                TimeUnit.MILLISECONDS
+              )
             case SCHEDULED =>
               shutdownSchedule = RESCHEDULED
             case RESCHEDULED => //Already marked for reschedule
           }
         }
       }
-  }
+    }
 
   private[akka] def register(actorRef: ActorRef) {
     if (actorRef.mailbox eq null) actorRef.mailbox = createMailbox(actorRef)
@@ -147,7 +163,10 @@ trait MessageDispatcher {
           case UNSCHEDULED =>
             shutdownSchedule = SCHEDULED
             Scheduler.scheduleOnce(
-                shutdownAction, timeoutMs, TimeUnit.MILLISECONDS)
+              shutdownAction,
+              timeoutMs,
+              TimeUnit.MILLISECONDS
+            )
           case SCHEDULED =>
             shutdownSchedule = RESCHEDULED
           case RESCHEDULED => //Already marked for reschedule
@@ -165,27 +184,28 @@ trait MessageDispatcher {
       val uuid = i.next()
       Actor.registry.actorFor(uuid) match {
         case Some(actor) => actor.stop()
-        case None => {}
+        case None        => {}
       }
     }
   }
 
   private val shutdownAction = new Runnable {
-    def run = guard withGuard {
-      shutdownSchedule match {
-        case RESCHEDULED =>
-          shutdownSchedule = SCHEDULED
-          Scheduler.scheduleOnce(this, timeoutMs, TimeUnit.MILLISECONDS)
-        case SCHEDULED =>
-          if (uuids.isEmpty && futures.get == 0) {
-            active switchOff {
-              shutdown // shut down in the dispatcher's references is zero
+    def run =
+      guard withGuard {
+        shutdownSchedule match {
+          case RESCHEDULED =>
+            shutdownSchedule = SCHEDULED
+            Scheduler.scheduleOnce(this, timeoutMs, TimeUnit.MILLISECONDS)
+          case SCHEDULED =>
+            if (uuids.isEmpty && futures.get == 0) {
+              active switchOff {
+                shutdown // shut down in the dispatcher's references is zero
+              }
             }
-          }
-          shutdownSchedule = UNSCHEDULED
-        case UNSCHEDULED => //Do nothing
+            shutdownSchedule = UNSCHEDULED
+          case UNSCHEDULED => //Do nothing
+        }
       }
-    }
   }
 
   /**
@@ -249,39 +269,51 @@ abstract class MessageDispatcherConfigurator {
     if (capacity < 1) UnboundedMailbox()
     else
       BoundedMailbox(
-          capacity,
-          Duration(
-              config.getInt("mailbox-push-timeout-time",
-                            Dispatchers.MAILBOX_PUSH_TIME_OUT.toMillis.toInt),
-              TIME_UNIT))
+        capacity,
+        Duration(
+          config.getInt(
+            "mailbox-push-timeout-time",
+            Dispatchers.MAILBOX_PUSH_TIME_OUT.toMillis.toInt
+          ),
+          TIME_UNIT
+        )
+      )
   }
 
   def configureThreadPool(
       config: Configuration,
-      createDispatcher: => (ThreadPoolConfig) => MessageDispatcher)
-    : ThreadPoolConfigDispatcherBuilder = {
+      createDispatcher: => (ThreadPoolConfig) => MessageDispatcher
+  ): ThreadPoolConfigDispatcherBuilder = {
     import ThreadPoolConfigDispatcherBuilder.conf_?
 
     //Apply the following options to the config if they are present in the config
     ThreadPoolConfigDispatcherBuilder(createDispatcher, ThreadPoolConfig())
-      .configure(conf_?(config getInt "keep-alive-time")(
-                     time => _.setKeepAliveTime(Duration(time, TIME_UNIT))),
-                 conf_?(config getDouble "core-pool-size-factor")(
-                     factor => _.setCorePoolSizeFromFactor(factor)),
-                 conf_?(config getDouble "max-pool-size-factor")(
-                     factor => _.setMaxPoolSizeFromFactor(factor)),
-                 conf_?(config getInt "executor-bounds")(
-                     bounds => _.setExecutorBounds(bounds)),
-                 conf_?(config getBool "allow-core-timeout")(
-                     allow => _.setAllowCoreThreadTimeout(allow)),
-                 conf_?(config getString "rejection-policy" map {
-                   case "abort" => new AbortPolicy()
-                   case "caller-runs" => new CallerRunsPolicy()
-                   case "discard-oldest" => new DiscardOldestPolicy()
-                   case "discard" => new DiscardPolicy()
-                   case x =>
-                     throw new IllegalArgumentException(
-                         "[%s] is not a valid rejectionPolicy!" format x)
-                 })(policy => _.setRejectionPolicy(policy)))
+      .configure(
+        conf_?(config getInt "keep-alive-time")(time =>
+          _.setKeepAliveTime(Duration(time, TIME_UNIT))
+        ),
+        conf_?(config getDouble "core-pool-size-factor")(factor =>
+          _.setCorePoolSizeFromFactor(factor)
+        ),
+        conf_?(config getDouble "max-pool-size-factor")(factor =>
+          _.setMaxPoolSizeFromFactor(factor)
+        ),
+        conf_?(config getInt "executor-bounds")(bounds =>
+          _.setExecutorBounds(bounds)
+        ),
+        conf_?(config getBool "allow-core-timeout")(allow =>
+          _.setAllowCoreThreadTimeout(allow)
+        ),
+        conf_?(config getString "rejection-policy" map {
+          case "abort"          => new AbortPolicy()
+          case "caller-runs"    => new CallerRunsPolicy()
+          case "discard-oldest" => new DiscardOldestPolicy()
+          case "discard"        => new DiscardPolicy()
+          case x =>
+            throw new IllegalArgumentException(
+              "[%s] is not a valid rejectionPolicy!" format x
+            )
+        })(policy => _.setRejectionPolicy(policy))
+      )
   }
 }
